reorg redux stuff into a dir like blip & uploader do + actions tests
