+++ package.json
@@ -67,6 +67,7 @@
     "karma-webpack": "1.8.0",
     "lodash": "3.10.1",
     "mocha": "3.0.2",
+    "moment-timezone": "0.5.5",
     "phantomjs-prebuilt": "2.1.12",
     "postcss": "5.1.2",
     "postcss-calc": "5.3.1",
@@ -82,19 +83,15 @@
     "rimraf": "2.5.4",
     "sinon": "1.17.5",
     "style-loader": "0.13.1",
+    "tidepool-standard-action": "0.1.1",
     "webpack": "1.13.2"
   },
   "peerDependencies": {
     "classnames": "2.x",
-    "d3-array": "1.x",
-    "d3-format": "1.x",
-    "d3-scale": "1.x",
-    "d3-shape": "1.x",
-    "d3-time": "1.x",
     "lodash": "3.x",
     "moment-timezone": "0.x",
+    "react": "15.x",
     "react-addons-update": "15.x",
-    "react-motion": "0.x",
     "react-redux": "4.x",
     "redux": "3.x"
   },
@@ -104,7 +101,6 @@
     "d3-scale": "1.0.3",
     "d3-shape": "1.0.3",
     "d3-time": "1.0.2",
-    "moment-timezone": "0.5.5",
     "react-motion": "0.4.4"
   }
 }
